+++ src/global_fnptr.hpp
@@ -35,9 +35,10 @@ namespace detail {
     friend class detail::command<FunctionTokenType, Arg...>;
     friend struct serialization<global_fnptr>;
 
+    constexpr global_fnptr(const FunctionToken& ft, detail::internal_only) : u_(ft) {}
+    constexpr global_fnptr(FunctionToken&& ft, detail::internal_only) : u_(std::move(ft)) {}
+
   private:
-    constexpr global_fnptr(const FunctionToken& ft) : u_(ft) {}
-    constexpr global_fnptr(FunctionToken&& ft) : u_(std::move(ft)) {}
     FunctionToken u_;
 
   public:
@@ -138,21 +139,26 @@ namespace detail {
       }
     }
 
-    template<typename Reader>
-    static detail::global_fnptr<Fn,detail::function_token>* deserialize(Reader& r, void* storage)
+    template<typename Reader, typename Storage>
+    static detail::global_fnptr<Fn,detail::function_token>* deserialize(Reader& r, Storage storage)
     {
       auto offset = r.template read_trivial<uintptr_t>();
       if (!(offset & msb)) {
         // issue 553: Workaround for CUDA 11.0.3 nvcc frontend bug where it fails to recognize the single-member aggregate initialization 
         // if the `function_token_ss` is constructed in the same line
         detail::function_token_ss ss{offset};
-        return ::new(storage) detail::global_fnptr<Fn,detail::function_token>{std::move(ss)};
+        return storage.construct(std::move(ss), detail::internal_only{});
       } else {
         int16_t idx = (offset & idx_bits) >> idx_shift;
         if (idx > 0)
-          return ::new(storage) detail::global_fnptr<Fn,detail::function_token>{detail::function_token_ms_idx{offset & lower_bits, idx}};
+          return storage.construct(detail::function_token_ms_idx{offset & lower_bits, idx},
+                                   detail::internal_only{});
         else
-          return ::new(storage) detail::global_fnptr<Fn,detail::function_token>{detail::function_token_ms{offset & lower_bits, r.template read_trivial<detail::segment_hash>()}};
+          return storage.construct(
+            detail::function_token_ms{offset & lower_bits,
+                                      r.template read_trivial<detail::segment_hash>()},
+            detail::internal_only{}
+          );
       }
     }
   };